requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
drf-spectacular>=0.27
celery>=5.4
psycopg[binary]>=3.2
mysqlclient>=2.2
gunicorn>=22.0
sentry-sdk>=1.50.0
prometheus-client>=0.20.0
python-json-logger>=2.0.7
